app/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2
django-mptt==0.11.0
mysqlclient

